pyproject.toml
[project]
name = "review-company-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.6",
    "mysql-connector-python>=9.2.0",
    "mysqlclient>=2.2.7",
    "pymysql>=1.1.1",
    "sqlalchemy>=2.0.37",
]
